refactor: Introduce `get-shared-stack-action-params` util
