+++ lib/plugins/aws/deploy/index.js
@@ -9,6 +9,7 @@ const checkForChanges = require('./lib/check-for-changes');
 const monitorStack = require('../lib/monitor-stack');
 const checkIfBucketExists = require('../lib/check-if-bucket-exists');
 const getCreateChangeSetParams = require('../lib/get-create-change-set-params');
+const getSharedStackActionParams = require('../lib/get-shared-stack-action-params');
 const getCreateStackParams = require('../lib/get-create-stack-params');
 const getUpdateStackParams = require('../lib/get-update-stack-params');
 const getExecuteChangeSetParams = require('../lib/get-execute-change-set-params');
@@ -56,7 +57,8 @@ class AwsDeploy {
       getCreateChangeSetParams,
       getCreateStackParams,
       getExecuteChangeSetParams,
-      getUpdateStackParams
+      getUpdateStackParams,
+      getSharedStackActionParams
     );
 
     this.getFileStats = memoize(this.getFileStats.bind(this), { promise: true });
