+++ lib/plugins/aws/rollback.js
@@ -7,6 +7,9 @@ const monitorStack = require('./lib/monitor-stack');
 const waitForChangeSetCreation = require('./lib/wait-for-change-set-creation');
 const getCreateChangeSetParams = require('./lib/get-create-change-set-params');
 const getExecuteChangeSetParams = require('./lib/get-execute-change-set-params');
+const getSharedStackActionParams = require('./lib/get-shared-stack-action-params');
+const getCreateStackParams = require('./lib/get-create-stack-params');
+const getUpdateStackParams = require('./lib/get-update-stack-params');
 const findAndGroupDeployments = require('./utils/find-and-group-deployments');
 const ServerlessError = require('../../serverless-error');
 const { style, log, progress } = require('@serverless/utils/log');
@@ -30,7 +33,10 @@ class AwsRollback {
       monitorStack,
       waitForChangeSetCreation,
       getCreateChangeSetParams,
-      getExecuteChangeSetParams
+      getExecuteChangeSetParams,
+      getCreateStackParams,
+      getUpdateStackParams,
+      getSharedStackActionParams
     );
 
     this.hooks = {
