+++ test/unit/lib/plugins/aws/deploy/index.test.js
@@ -342,6 +342,7 @@ describe('test/unit/lib/plugins/aws/deploy/index.test.js', () => {
         StackName: awsNaming.getStackName(),
         Capabilities: ['CAPABILITY_IAM', 'CAPABILITY_NAMED_IAM'],
         Parameters: [],
+        NotificationARNs: [],
         Tags: [{ Key: 'STAGE', Value: 'dev' }],
         TemplateBody: JSON.stringify({
           Resources: serverless.service.provider.coreCloudFormationTemplate.Resources,
